feat: implement Portfolio Unit Investment calculation mode
- Add portfolioUnitInvestment enum case to CalculationMode (iOS/Android)
- Extend SavedCalculation models with portfolio unit investment parameters:
  * unitPrice: Price per unit/lead/patent/etc.
  * successRate: Percentage of units that succeed (0-100)
  * outcomePerUnit: Revenue/settlement/profit per successful unit
  * investorShare: Percentage of outcome that goes to investor (0-100)
  * feePercentage: Management/legal/servicing fees percentage (0-100)
- Add comprehensive validation logic for portfolio unit investment parameters
- Implement portfolio unit investment calculation methods in IRRCalculator:
  * calculatePortfolioUnitIRR(): Basic portfolio unit investment IRR
  * calculatePortfolioUnitBlendedIRR(): Multiple batches with different prices/timing
  * portfolioUnitGrowthPoints(): Growth chart data generation
- Add PortfolioUnitBatch data structure for follow-on investment batches
- Update UI components with placeholder views for new calculation mode
- Add validation and error handling for portfolio unit investment inputs

Supports litigation financing, patent portfolios, debt collection, royalty investments, and other unit-based investment scenarios with success rates and investor share modeling.

Tasks completed: 7.1, 7.2 (Portfolio Unit Investment enum/data model and calculation logic)

+++ android/app/src/main/java/com/irrgenius/android/ui/components/GrowthChartView.kt
@@ -91,10 +91,10 @@ private fun SimpleLineChart(
         val height = size.height - 2 * padding
         
         // Find min/max values
-        val minX = points.minOf { it.month }
-        val maxX = points.maxOf { it.month }
-        val minY = points.minOf { it.value }
-        val maxY = points.maxOf { it.value }
+        val minX = points.minOf { it.month }.toFloat()
+        val maxX = points.maxOf { it.month }.toFloat()
+        val minY = points.minOf { it.value }.toFloat()
+        val maxY = points.maxOf { it.value }.toFloat()
         
         val xRange = maxX - minX
         val yRange = maxY - minY
@@ -107,8 +107,8 @@ private fun SimpleLineChart(
         // Create path for line
         val path = Path()
         points.forEachIndexed { index, point ->
-            val x = padding + ((point.month - minX) / xRange) * width
-            val y = padding + height - ((point.value - minY) / yRange) * height
+            val x = padding + ((point.month.toFloat() - minX) / xRange) * width
+            val y = padding + height - ((point.value.toFloat() - minY) / yRange) * height
             
             if (index == 0) {
                 path.moveTo(x, y)
@@ -126,8 +126,8 @@ private fun SimpleLineChart(
         
         // Draw points
         points.forEach { point ->
-            val x = padding + ((point.month - minX) / xRange) * width
-            val y = padding + height - ((point.value - minY) / yRange) * height
+            val x = padding + ((point.month.toFloat() - minX) / xRange) * width
+            val y = padding + height - ((point.value.toFloat() - minY) / yRange) * height
             
             drawCircle(
                 color = lineColor,
